requirements.txt
alembic==1.4.3
aniso8601==8.1.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
flask-serialize==1.4.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
permissive-dict==1.0.3
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-serializer==1.3.4.4
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-JSON==0.3.3
